Filter circles by location; send AJAX post request upon clicking join request button

+++ lib/run_pal/use_cases/circle_location_filter.rb
@@ -2,8 +2,8 @@ module RunPal
   class FilterCirclesByLocation < UseCase
 
     def run(inputs)
-      inputs[:user_lat] = inputs[:user_lat] ? inputs[:user_lat].to_i : nil
-      inputs[:user_long] = inputs[:user_long] ? inputs[:user_long].to_i : nil
+      inputs[:user_lat] = inputs[:user_lat] ? inputs[:user_lat].to_f : nil
+      inputs[:user_long] = inputs[:user_long] ? inputs[:user_long].to_f : nil
       inputs[:radius] = inputs[:radius] ? inputs[:radius].to_i : nil
 
       circles = get_nearby_circles(inputs)
